requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
idna==3.7
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
